refactor(contexts): tidy things around
