+++ src/hooks/use-process-query.ts
@@ -2,7 +2,7 @@ import { useQuery } from '@tanstack/react-query'
 import type { ElectionMetadata } from '@vocdoni/davinci-sdk'
 import { useEffect } from 'react'
 import { up } from 'up-fetch'
-import { useVocdoniApi } from '~components/vocdoni-api-context'
+import { useVocdoniApi } from '~contexts/vocdoni-api-context'
 import type { Process } from '~src/types'
 
 const upfetch = up(fetch)
