+++ src/main.tsx
@@ -5,9 +5,9 @@ import { AppErrorBoundary } from '~components/app-error-boundary'
 import { InstallPrompt } from '~components/install-prompt'
 import { OfflineIndicator } from '~components/offline-indicator'
 import { UpdatePrompt } from '~components/update-prompt'
-import { VocdoniApiProvider } from '~components/vocdoni-api-context'
 import { MiniAppProvider } from '~contexts/MiniAppContext'
 import { SequencerNetworkProvider } from '~contexts/sequencer-network'
+import { VocdoniApiProvider } from '~contexts/vocdoni-api-context'
 import { RouterProvider } from '~router'
 // Initialize AppKit with all networks
 import '~lib/appkit'
